Fix calculo con variables negativas. Cambios para mejorar performance

+++ Generadores/ChiCuadrado.cs
@@ -5,17 +5,17 @@ namespace NumerosAleatorios
 {
     public static class ChiCuadrado
     {
-        public static double Calcular(double observada, double esperada)
+        public static float Calcular(float observada, float esperada)
         {
             if (!(esperada > 0))
                 throw new NotSupportedException("La frecuencia esperada debe ser mayor a cero");
 
-            var valor = Math.Pow(esperada - observada, 2) / esperada;
+            var valor = (float)Math.Pow(esperada - observada, 2) / esperada;
 
             return valor;
         }
 
-        public static double ValorDeTabla(int grados, double alfa)
+        public static float ValorDeTabla(int grados, float alfa)
         {
             if (!(grados > 0))
                 throw new NotSupportedException("Los grados de libertad deben ser mayores a cero");
@@ -23,7 +23,7 @@ namespace NumerosAleatorios
             var d = new ChiSquaredDistribution(grados);
 
             var valor = d.InverseRightProbability(alfa);
-            return valor;
+            return (float)valor;
         }
     }
 }
