+++ Generadores/Distribuciones/IDistribucion.cs
@@ -6,9 +6,9 @@ namespace NumerosAleatorios.VariablesAleatorias
     public interface IDistribucion
     {
         void AsignarGenerador(IGeneradorNumerosAleatorios generador);
-        double Generar();
-        List<double> Generar(int cantidad);
-        List<double> FrecuenciasEsperadas(List<Intervalo> intervalos);
+        float Generar();
+        List<float> Generar(int cantidad);
+        List<float> FrecuenciasEsperadas(List<Intervalo> intervalos);
         int CantidadParametros();
     }
 }
