+++ Generadores/Generacion NumerosAleatorios/CongruencialMultiplicativo.cs
@@ -4,18 +4,18 @@ namespace NumerosAleatorios.NumerosAleatorios
 {
     public class CongruencialMultiplicativo : IGeneradorNumerosAleatorios
     {
-        public double Semilla { get; protected set; }
-        public double A { get; protected set; }
-        public double M { get; protected set; }
+        public float Semilla { get; protected set; }
+        public float A { get; protected set; }
+        public float M { get; protected set; }
 
-        public CongruencialMultiplicativo(double semilla, double a, double m)
+        public CongruencialMultiplicativo(float semilla, float a, float m)
         {
             Semilla = semilla;
             A = a;
             M = m;
         }
 
-        public double Generar()
+        public float Generar()
         {
             var aleatorio = (A * Semilla) % M;
